requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
supabase==2.5.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.2.1
# test/dev
pytest

# HR / OCR / Ranking
pdfplumber==0.11.0
python-docx==1.1.2
groq==0.11.0
pytz==2024.1
